+++ src/main/java/ru/javawebinar/topjava/service/UserService.java
@@ -9,9 +9,9 @@ public interface UserService {
 
     User create(User user);
 
-    void delete(int id) throws NotFoundException;
+    void delete(Long id) throws NotFoundException;
 
-    User get(int id) throws NotFoundException;
+    User get(Long id) throws NotFoundException;
 
     User getByEmail(String email) throws NotFoundException;
