+++ src/main/java/ru/javawebinar/topjava/service/impl/UserServiceImpl.java
@@ -14,12 +14,12 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void delete(int id) throws NotFoundException {
+    public void delete(Long id) throws NotFoundException {
 
     }
 
     @Override
-    public User get(int id) throws NotFoundException {
+    public User get(Long id) throws NotFoundException {
         return null;
     }
